+++ MASStorage/plugins/cordova-plugin-mas-storage/src/android/MASStoragePlugin.java
@@ -14,7 +14,6 @@ import com.ca.mas.foundation.MASCallback;
 import com.ca.mas.foundation.MASConstants;
 import com.ca.mas.storage.DataMarshaller;
 import com.ca.mas.storage.MASSecureLocalStorage;
-import com.ca.mas.storage.MASSecureStorage;
 import com.ca.mas.storage.MASStorage;
 import com.ca.mas.storage.MASStorageSegment;
 
@@ -33,10 +32,6 @@ public class MASStoragePlugin extends MASCordovaPlugin {
     private static final int CORDOVA_MAS_LOCAL_STORAGE_SEGMENT_APPLICATION = 0;
     private static final int CORDOVA_MAS_LOCAL_STORAGE_SEGMENT_APPLICATION_FOR_USER = 1;
 
-    private static final int CORDOVA_MAS_CLOUD_STORAGE_SEGMENT_USER = 0;
-    private static final int CORDOVA_MAS_CLOUD_STORAGE_SEGMENT_APPLICATION = 1;
-    private static final int CORDOVA_MAS_CLOUD_STORAGE_SEGMENT_APPLICATION_FOR_USER = 2;
-
     @Override
     protected void pluginInitialize() {
         super.pluginInitialize();
@@ -54,14 +49,6 @@ public class MASStoragePlugin extends MASCordovaPlugin {
             deleteAllUsingModeLocal(args, callbackContext);
         } else if (action.equalsIgnoreCase("keySetForModeLocal")) {
             keySetForModeLocal(args, callbackContext);
-        } else if (action.equalsIgnoreCase("saveToCloud")) {
-            saveToCloud(args, callbackContext);
-        } else if (action.equalsIgnoreCase("findByUsingKeyAndModeCloud")) {
-            findByUsingKeyAndModeCloud(args, callbackContext);
-        } else if (action.equalsIgnoreCase("deleteByUsingKeyAndModeCloud")) {
-            deleteByUsingKeyAndModeCloud(args, callbackContext);
-        } else if (action.equalsIgnoreCase("keySetForModeCloud")) {
-            keySetForModeCloud(args, callbackContext);
         } else {
             callbackContext.error("Invalid action");
             return false;
@@ -199,128 +186,6 @@ public class MASStoragePlugin extends MASCordovaPlugin {
         }
     }
 
-    private void saveToCloud(final JSONArray args, final CallbackContext callbackContext) {
-        try {
-            MASStorage storage = new MASSecureStorage();
-            String key = args.optString(1);
-            Object data = args.opt(2);
-            int segment_0 = args.getInt(3);
-            int segment = fetchSegmentCloud(segment_0);
-
-            storage.save(key, data, segment, new MASCallback<Void>() {
-                @Override
-                public void onSuccess(Void result) {
-                    success(callbackContext, true, false);
-                }
-
-                @Override
-                public void onError(Throwable e) {
-                    Log.e(TAG, e.getMessage());
-                    callbackContext.error(getError(e));
-                }
-            });
-        } catch (Exception e) {
-            Log.e(TAG, e.getMessage(), e);
-            callbackContext.error(getError(e));
-        }
-    }
-
-    private void findByUsingKeyAndModeCloud(final JSONArray args, final CallbackContext callbackContext) {
-        try {
-            MASStorage storage = new MASSecureStorage();
-            String key = args.optString(0);
-            int segment_0 = args.getInt(1);
-            int segment = fetchSegmentCloud(segment_0);
-
-            storage.findByKey(key, segment, new MASCallback() {
-                @Override
-                public void onSuccess(Object result) {
-                    JSONObject response = null;
-                    try {
-                        response = getResultJson(result);
-                        success(callbackContext, response, false);
-                    } catch (Exception ex) {
-                        LOG.e(TAG, ex.getMessage());
-                        callbackContext.error(getError(ex));
-                    }
-                }
-
-                @Override
-                public void onError(Throwable e) {
-                    LOG.e(TAG, e.getMessage());
-                    callbackContext.error(getError(e));
-                }
-            });
-        } catch (Exception e) {
-            Log.e(TAG, e.getMessage(), e);
-            callbackContext.error(getError(e));
-        }
-    }
-
-    private void deleteByUsingKeyAndModeCloud(final JSONArray args, final CallbackContext callbackContext) {
-        try {
-            MASStorage storage = new MASSecureStorage();
-            String key = args.optString(0);
-            int segment_0 = args.getInt(1);
-            int segment = fetchSegmentCloud(segment_0);
-
-            storage.delete(key, segment, new MASCallback<Void>() {
-                @Override
-                public void onSuccess(Void result) {
-                    success(callbackContext, true, false);
-                }
-
-                @Override
-                public void onError(Throwable e) {
-                    Log.e(TAG, e.getMessage(), e);
-                    callbackContext.error(getError(e));
-                }
-            });
-        } catch (Exception e) {
-            Log.e(TAG, e.getMessage(), e);
-            callbackContext.error(getError(e));
-        }
-    }
-
-    private void keySetForModeCloud(final JSONArray args, final CallbackContext callbackContext) {
-        try {
-            MASStorage storage = new MASSecureStorage();
-            int segment_0 = args.getInt(0);
-            int segment = fetchSegmentCloud(segment_0);
-
-            storage.keySet(segment, new MASCallback<Set<String>>() {
-                @Override
-                public void onSuccess(Set<String> result) {
-                    JSONArray array = new JSONArray(result);
-                    success(callbackContext, array, false);
-                }
-
-                @Override
-                public void onError(Throwable e) {
-                    Log.e(TAG, e.getMessage(), e);
-                    callbackContext.error(getError(e));
-                }
-            });
-        } catch (Exception e) {
-            Log.e(TAG, e.getMessage(), e);
-            callbackContext.error(getError(e));
-        }
-    }
-
-    private
-    @MASStorageSegment
-    int fetchSegmentCloud(int segment) {
-        switch (segment) {
-            case CORDOVA_MAS_CLOUD_STORAGE_SEGMENT_USER:
-                return MASConstants.MAS_USER;
-            case CORDOVA_MAS_CLOUD_STORAGE_SEGMENT_APPLICATION:
-                return MASConstants.MAS_APPLICATION;
-            case CORDOVA_MAS_CLOUD_STORAGE_SEGMENT_APPLICATION_FOR_USER:
-                return MASConstants.MAS_USER | MASConstants.MAS_APPLICATION;
-            default:
-                throw new UnsupportedOperationException("This segment is not mapped to any of the present segments");
-        }
-    }
 
     private static
     @MASStorageSegment